+++ src/DDTrace/Integrations/PDO/PDOIntegration.php
@@ -31,6 +31,15 @@ class PDOIntegration
             return Integration::NOT_AVAILABLE;
         }
 
+        /**
+         * Workaround for the Drupal DBAL bug.
+         * Delete this `if` block once the bug is fixed.
+         * @see https://github.com/DataDog/dd-trace-php/pull/284
+         */
+        if (defined('DRUPAL_CORE_COMPATIBILITY')) {
+            return Integration::NOT_AVAILABLE;
+        }
+
         // public PDO::__construct ( string $dsn [, string $username [, string $passwd [, array $options ]]] )
         dd_trace('PDO', '__construct', function () {
             $args = func_get_args();