+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file - [read more
 
 ### Fixed
 - Symfony 4.2 traces generation #280
+- Drupal crashes (temporary workaround) #285
 
 ## [0.12.2]